1570 - separated AiCopilotFacadeImpl into 2 files
