+++ server/libs/platform/platform-ai/platform-ai-service/src/main/java/com/bytechef/platform/ai/repository/converter/MapToPGObjectConverter.java
@@ -18,13 +18,16 @@ package com.bytechef.platform.ai.repository.converter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.sql.SQLException;
+import java.util.HashMap;
 import java.util.Map;
 import org.postgresql.util.PGobject;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.WritingConverter;
 
 @WritingConverter
+@SuppressFBWarnings("EI_EXPOSE_REP2")
 public class MapToPGObjectConverter implements Converter<Map<String, Object>, PGobject> {
     private final ObjectMapper objectMapper;
 
@@ -34,13 +37,16 @@ public class MapToPGObjectConverter implements Converter<Map<String, Object>, PG
 
     @Override
     public PGobject convert(Map<String, Object> source) {
+        Map<String, Object> defensiveCopy = new HashMap<>(source);
         try {
             PGobject jsonObject = new PGobject();
             jsonObject.setType("jsonb");
-            jsonObject.setValue(objectMapper.writeValueAsString(source));
-            return jsonObject;
+            jsonObject.setValue(objectMapper.writeValueAsString(defensiveCopy));
+            return (PGobject) jsonObject.clone();
         } catch (SQLException | JsonProcessingException e) {
             throw new RuntimeException("Failed to convert Map to PGobject", e);
+        } catch (CloneNotSupportedException e) {
+            throw new RuntimeException("Could not clone PGobject", e);
         }
     }
 }
